Fix reading Jicamarca data files by detecting the end of file properly.
Numpy's fromfile returns a zero-length array if the end of the file is reached. We were
accessing the data using the item() method with no argument and checking for an
IndexError. In fact, numpy raises and ValueError when the array is zero-length and item
is called with no arguments. This replaces all IndexError checks in that situation with
ValueError checks.